+++ app/schemas/sale.py
@@ -1,6 +1,5 @@
 import uuid
 from datetime import date, datetime
-from decimal import Decimal
 from typing import Annotated
 
 from pydantic import BaseModel, Field
@@ -10,7 +9,7 @@ from app.models.enums import SaleStatus
 
 class SaleItemInput(BaseModel):
     product_id: uuid.UUID
-    quantity: Annotated[Decimal, Field(gt=0, description="Must be > 0")]
+    quantity: Annotated[float, Field(gt=0, description="Must be > 0")]
 
 
 class SaleCreate(BaseModel):
@@ -23,17 +22,17 @@ class SaleItemResponse(BaseModel):
     id: uuid.UUID
     product_id: uuid.UUID
     product_name: str
-    quantity: Decimal
-    sale_price: Decimal
-    purchase_price: Decimal
+    quantity: float
+    sale_price: float
+    purchase_price: float
 
 
 class SaleResponse(BaseModel):
     id: uuid.UUID
     location_id: uuid.UUID
     status: SaleStatus
-    total_amount: Decimal
-    total_cost: Decimal
+    total_amount: float
+    total_cost: float
     note: str | None
     created_at: datetime
     items: list[SaleItemResponse] = []
@@ -49,11 +48,11 @@ class SaleListResponse(BaseModel):
 
 
 class SalesSummary(BaseModel):
-    total_revenue: Decimal
-    total_cost: Decimal
-    total_profit: Decimal
-    profit_margin: Decimal
+    total_revenue: float
+    total_cost: float
+    total_profit: float
+    profit_margin: float
     sales_count: int
-    average_sale: Decimal
+    average_sale: float
     date_from: date | None = None
     date_to: date | None = None
